Add getStaticProps for categories
