+++ queries/product-by-category.js
@@ -1,7 +1,7 @@
 import { gql } from "@apollo/client";
 
 const PRODUCT_BY_CATEGORY_ID = gql` query Product_Category($id: ID!) {
-	productCategory(id: $id) {
+	productCategory(id: $id, idType: SLUG) {
 	  id
 	  name
 	  products(first: 50) {
